requirements.txt
django-debug-toolbar==3.2.2
django-extensions==3.1.3
Django==3.2.6
